pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pyautolab"
description = "A python application for scientific experiments automation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT License"}
version = "0.1.0"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Environment :: X11 Applications :: Qt",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "psutil",
    "pyserial",
    "numpy",
    "Qtpy",
    "QtAwesome",
    "pyqtgraph",
    "pyqtdarktheme",
]

[tool.setuptools.package-data]
"*" = ["*.json"]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pyinstaller",
    # flake8
    "flake8",
    "flake8-return",
    "flake8-pie",
    "flake8-print",
    "flake8-docstrings",
    "flake8-bugbear",
    "flake8-simplify",
    "flake8-comprehensions",
    "flake8-eradicate",
    "flake8-rst-docstrings",
    "flake8-pyproject",
    "flake8-pytest-style",
    "pep8-naming",
]

[project.gui-scripts]
pyautolab = "pyautolab.__main__:main"

[project.entry-points."pyinstaller40"]
hook-dirs = "pyautolab.__pyinstaller:get_hook_dirs"

[tool.flake8]
max-line-length = 119
max-complexity = 13
extend-ignore = "E203"
per-file-ignores = ["**/__init__.py:F401"]

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119
